requirements.txt
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
Django==5.1.4
django-allauth==65.3.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
redis==5.2.1
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
ipython==8.30.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
Pillow==10.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
